+++ CHANGELOG.md
@@ -10,7 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 
 * `union()`
-* `convertCase` option to `string()`
+* `caseConverter` option to `string()`
 
 ### Changed